requirements.txt
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
folium==0.19.7
geopandas==1.1.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pandas==2.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyogrio==0.11.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
shapely==2.1.1
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
xyzservices==2025.4.0
